CheckpointStore (#51)
* WIP

* CheckpointStore

* rocksdb stuff

* assorted renaming

* move RpcClientExtensions back

* changelog

* update changelog

* RocksDbStoreTest

* add storage tests

* change test assembly namespace

* Add empty key + backwards checks to other Seek implemenations

* dotnet format

* add checkpoint metadata test

+++ src/bctklib/persistence/RocksDbStore.cs
@@ -16,12 +16,12 @@ namespace Neo.BlockchainToolkit.Persistence
         readonly WriteOptions writeOptions = new WriteOptions();
         readonly WriteOptions writeSyncOptions = new WriteOptions().SetSync(true);
 
-        public RocksDbStore(RocksDb db, string? columnFamilyName = null, bool readOnly = false)
+        internal RocksDbStore(RocksDb db, string? columnFamilyName = null, bool readOnly = false)
             : this(db, GetColumnFamilyHandle(db, columnFamilyName), readOnly, false)
         {
         }
 
-        public RocksDbStore(RocksDb db, ColumnFamilyHandle columnFamily, bool readOnly = false)
+        internal RocksDbStore(RocksDb db, ColumnFamilyHandle columnFamily, bool readOnly = false)
             : this(db, columnFamily, readOnly, false)
         {
         }
@@ -48,46 +48,47 @@ namespace Neo.BlockchainToolkit.Persistence
 
         public byte[]? TryGet(byte[]? key)
         {
-            if (disposed) throw new ObjectDisposedException(nameof(RocksDbStore));
+            if (disposed || db.Handle == IntPtr.Zero) throw new ObjectDisposedException(nameof(RocksDbStore));
             return db.Get(key ?? Array.Empty<byte>(), columnFamily, readOptions);
         }
 
         public bool Contains(byte[]? key)
         {
-            if (disposed) throw new ObjectDisposedException(nameof(RocksDbStore));
+            if (disposed || db.Handle == IntPtr.Zero) throw new ObjectDisposedException(nameof(RocksDbStore));
             return TryGet(key) != null;
         }
 
         public IEnumerable<(byte[] Key, byte[] Value)> Seek(byte[]? key, SeekDirection direction)
         {
-            if (disposed) throw new ObjectDisposedException(nameof(RocksDbStore));
+            if (disposed || db.Handle == IntPtr.Zero) throw new ObjectDisposedException(nameof(RocksDbStore));
             return db.Seek(columnFamily, key, direction, readOptions);
         }
 
         public void Put(byte[]? key, byte[] value)
         {
-            if (disposed) throw new ObjectDisposedException(nameof(RocksDbStore));
+            if (disposed || db.Handle == IntPtr.Zero) throw new ObjectDisposedException(nameof(RocksDbStore));
             if (readOnly) throw new InvalidOperationException("read only");
             db.Put(key ?? Array.Empty<byte>(), value, columnFamily, writeOptions);
         }
 
         public void PutSync(byte[]? key, byte[] value)
         {
-            if (disposed) throw new ObjectDisposedException(nameof(RocksDbStore));
+            if (disposed || db.Handle == IntPtr.Zero) throw new ObjectDisposedException(nameof(RocksDbStore));
             if (readOnly) throw new InvalidOperationException("read only");
             db.Put(key ?? Array.Empty<byte>(), value, columnFamily, writeSyncOptions);
         }
 
         public void Delete(byte[]? key)
         {
-            if (disposed) throw new ObjectDisposedException(nameof(RocksDbStore));
+            if (disposed || db.Handle == IntPtr.Zero) throw new ObjectDisposedException(nameof(RocksDbStore));
             if (readOnly) throw new InvalidOperationException("read only");
             db.Remove(key ?? Array.Empty<byte>(), columnFamily, writeOptions);
         }
 
         public ISnapshot GetSnapshot()
         {
-            if (disposed) throw new ObjectDisposedException(nameof(RocksDbStore));
+            if (disposed || db.Handle == IntPtr.Zero) throw new ObjectDisposedException(nameof(RocksDbStore));
+            if (readOnly) throw new InvalidOperationException("read only");
             return new Snapshot(db, columnFamily);
         }
     }
